Remove wildcard domains because of TransIP API bug
Due to an unknown bug in either the Ruby client, or the TransIP API, DNS entries
with wildcards (*) in them, will fail the request:

  Transip::ApiError: Invalid API signature, signature does not match the
  request. (timestamp: 0.74032400 1412434601)

These domains are therefor filtered out (and thus no longer work). This can be
problematic, but until a workaround is possible, this is the only way to work
with DNS entries.